Introduce support for a slabbed allocator:
When instantiating a large amount of fixed-sized objects on the heap
the overhead that dynamic memory allocation APIs impose will quickly
become significant.

In some cases, allocating a large amount of memory at once and using
a slabbing allocator to carve the large block into fixed-sized units
that are used to service requests for memory out will help to reduce
memory fragmentation significantly and, potentially, improve overall
performance.

This commit introduces a new SlabAllocator<> class that wraps around
the complexity and exposes an API that is similar to the C++ concept
of an Allocator.

This class is not a general-purpose allocator and should not be used
unless profiling and analysis of specific memory allocation patterns
indicates that the additional complexity introduced will improve the
performance of the system overall and subsequenting profiling proves
that.